live_logreader to its own file (#31436)
own file
